Move the "show more" button into a separate component
